+++ java/org/apache/tomcat/websocket/PerMessageDeflate.java
@@ -94,7 +94,7 @@ public class PerMessageDeflate implements Transformation {
                             throw new IllegalArgumentException(sm.getString("perMessageDeflate.invalidWindowSize",
                                     SERVER_MAX_WINDOW_BITS, Integer.valueOf(serverMaxWindowBits)));
                         }
-                        // Java SE API (as of Java 8) does not expose the API to
+                        // Java SE API (as of Java 11) does not expose the API to
                         // control the Window size. It is effectively hard-coded
                         // to 15
                         if (isServer && serverMaxWindowBits != 15) {
@@ -114,7 +114,7 @@ public class PerMessageDeflate implements Transformation {
                     if (clientMaxWindowBits == -1) {
                         if (param.getValue() == null) {
                             // Hint to server that the client supports this
-                            // option. Java SE API (as of Java 8) does not
+                            // option. Java SE API (as of Java 11) does not
                             // expose the API to control the Window size. It is
                             // effectively hard-coded to 15
                             clientMaxWindowBits = 15;
@@ -125,7 +125,7 @@ public class PerMessageDeflate implements Transformation {
                                         CLIENT_MAX_WINDOW_BITS, Integer.valueOf(clientMaxWindowBits)));
                             }
                         }
-                        // Java SE API (as of Java 8) does not expose the API to
+                        // Java SE API (as of Java 11) does not expose the API to
                         // control the Window size. It is effectively hard-coded
                         // to 15
                         if (!isServer && clientMaxWindowBits != 15) {
